Development

+++ TeamWorkFlow/Controllers/OperatorController.cs
@@ -1,5 +1,7 @@
 ﻿using Microsoft.AspNetCore.Mvc;
+using NuGet.Packaging.Signing;
 using TeamWorkFlow.Core.Contracts;
+using TeamWorkFlow.Core.Extensions;
 using TeamWorkFlow.Core.Models.Operator;
 using static TeamWorkFlow.Core.Constants.Messages;
 
@@ -61,9 +63,14 @@ namespace TeamWorkFlow.Controllers
         }
 
         [HttpGet]
-        public async Task<IActionResult> Edit(int id)
+        public async Task<IActionResult> Edit(int id, string extension)
         {
-	        if (ModelState.IsValid == false)
+            if (await _operatorService.OperatorExistByIdAsync(id) == false)
+            {
+                return BadRequest();
+            }
+
+            if (ModelState.IsValid == false)
 	        {
 		        return BadRequest();
 	        }
@@ -73,6 +80,11 @@ namespace TeamWorkFlow.Controllers
 			{
 				model.AvailabilityStatusModels = await _operatorService.GetAllOperatorStatusesAsync();
 			}
+
+            if (extension != model?.GetOperatorExtension())
+            {
+                return BadRequest();
+            }
 			
 			return View(model);
         }
@@ -104,41 +116,66 @@ namespace TeamWorkFlow.Controllers
 
             await _operatorService.EditOperatorAsync(model, id);
 
-            return RedirectToAction(nameof(All));
+            return RedirectToAction(nameof(Details), new {id = model.Id, extension = model.GetOperatorExtension()});
         }
 
 		[HttpGet]
-        public async Task<IActionResult> Details(int id)
+        public async Task<IActionResult> Details(int id, string extension)
         {
+            if (await _operatorService.OperatorExistByIdAsync(id) == false)
+            {
+                return BadRequest();
+            }
+
 	        if (!ModelState.IsValid)
 	        {
-		        BadRequest();
+		        return BadRequest();
 	        }
 
 	        var operatorModel = await _operatorService.GetOperatorDetailsByIdAsync(id);
+
+            if (extension != operatorModel?.GetOperatorExtension())
+            {
+                return BadRequest();
+            }
 
 	        return View(operatorModel);
         }
 
 		[HttpGet]
-        public async Task<IActionResult> Delete(int id)
+        public async Task<IActionResult> Delete(int id, string extension)
         {
-	        if (!ModelState.IsValid)
+            if (await _operatorService.OperatorExistByIdAsync(id) == false)
+            {
+                return BadRequest();
+            }
+
+            if (!ModelState.IsValid)
 	        {
-		        BadRequest();
+		        return BadRequest();
 	        }
 
 	        var operatorModel = await _operatorService.GetOperatorModelForDeleteByIdAsync(id);
 
-			return View(operatorModel);
+            if (extension != operatorModel?.GetOperatorExtension())
+            {
+                return BadRequest();
+            }
+
+            return View(operatorModel);
 		}
 
 		[HttpPost]
         public async Task<IActionResult> DeleteConfirmation(int id)
         {
+            if (await _operatorService.OperatorExistByIdAsync(id) == false)
+            {
+                return BadRequest();
+            }
+
 			if (!ModelState.IsValid)
 			{
-				BadRequest();
+				return BadRequest();
 			}
 
 			await _operatorService.DeleteOperatorByIdAsync(id);
